update and move files
