+++ pyproject.toml
@@ -13,9 +13,9 @@ dask = {extras = ["array"], version = "^2022.11.1"}
 ase = "^3.22.1"
 
 [tool.poetry.group.dev.dependencies]
-black = "^22.10.0"
-isort = "^5.10.1"
-ruff = "^0.0.149"
+black = "^23.1.0"
+isort = "^5.12.0"
+ruff = "^0.0.256"
 pytest = "^7.2.0"
 graphviz = "^0.20.1"
 pre-commit = "^2.20.0"
