+++ znh5md/format/__init__.py
@@ -1,3 +1,4 @@
+"""Format handler for h5md files."""
 import dataclasses
 import typing
